ui: stop relying on VSYNC for 60fps
We now support proper frame synchronization, and handle all that
internally instead of being a side-effect of vsync being enabled in the
UI.

This is a requirement for having the PSG work properly. Otherwise missed
frames (during VDP Blank time) will completely botch the PSG
synchronization. Now we can take them into account and simply not
refresh the display.