ui: add web support
This follows pixels' minimal-web example. Pixels building needs to be
async, probably from the way wgpu is initialized:
 * main function code is moved to an async function that is waited-on
   with the pollster crate on native, and spawn with
   wasm_bindgen_futures on web
 * Pixels is built with build_async() instead of build()
 * For debugging convenience, the panics are hooked to the console.

Since we don't use logging via the log crate with console_log plugging,
nothing else will be sent to the javascript console. We will find a
solution for that later.

We fix wasm-bindgen version, as this needs to be the same as the
wasm-bindgen-cli installed on the host. Currently we build like this:

  env RUSTFLAGS=--cfg=web_sys_unstable_apis cargo build --release  --target wasm32-unknown-unknown && wasm-bindgen --out-dir web --web ../target/wasm32-unknown-unknown/release/gears-ui.wasm

The minimal index.htm is provided.

No audio support yet, and it panics on thread::sleep().