psg: replace mpsc with custom channel but keep semantics
We don't need much, we don't even need mpsc, just spsc. But mpsc channel
don't work on wasm targets because the use Condvar to tell the OS they
are waiting, which is not supported on wasm. Mutex will have to do.

We actually go back to the previous implementation, with one exception:
instead of locking the cmd queue once and then clearing it, we are
instead locking it before getting every command, thus reducing the time
it's spending locked.