psg: ignore unused PSG data write frequency bit 6
This fixes sverx's PSGlib-based audio play (GGtest, Monaco Master...)
that uses this bit to signal something to send to the PSG (as opposed to
command options). It is of no consequence for attenuation (tone&noise)
and noise ctrl, but for frequency data we forgot to ignore this bit.

This lead to impossible 11-bits frequencies (counter values) that were
very low (< 100Hz).