ui: make core module public
Otherwise when building the lib outside of the wasm32 target (the only
place where it is used), it will seem like the code is unused.

Since we want the ui library to be a cdylib in order to launch the
entrypoint directly, we can't have the main depend on it.

This is purely to prevent splitting the UI into two crates (although we
could do this in the current workspace setup): one for the core library
and main, and another for the web library (cdylib).