psg: add low-pass filter on output
This biquad filter cites its sources in the code, in particular the
webaudio tutorial and wikipedia were very helpful. The biquad crate
provides the same functionality with no runtime overhead, but wasn't
necessary to import here.

The cutoff frequency of 2.7kHz is probably a bit too low, but provided a
good middle ground.

This version is far from perfect, but it improves audio slightly. A
better approch would be to let nyquist do its magic and render all audio
at 22kHz, so no frequency above 11kHz can be heard.